requirements.txt
awesome-slugify==1.6.5
Django==2.0.3
django-js-asset==1.0.0
Pillow==5.0.0
pytz==2018.3
regex==2018.2.21
Unidecode==0.4.21
